+++ yivi_app/lib/main.dart
@@ -5,6 +5,7 @@ import "package:yivi_core/yivi_core.dart";
 
 import "ocr_processor.dart";
 import "qr_scanner_factory.dart";
+import "regula_face_service.dart";
 import "sms_retriever.dart";
 
 void main() {
@@ -14,5 +15,6 @@ void main() {
     smsRetriever: Platform.isAndroid
         ? SmartAuthSmsRetriever(SmartAuth.instance)
         : null,
+    regulaFaceService: RegulaFaceServiceImpl(),
   );
 }